dockers/binance_features/requirements.txt
python-binance==1.0.16
pandas==1.3.4
numpy==1.20.3
argparse==1.4.0
psycopg2-binary==2.9.6
sqlalchemy==1.4.46
google-cloud-storage==2.5.0
